+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce new `formatter:` options: `:json` and `:passthrough`
 - Introduce new `transform:` options: `:cloud_watch` and `:passthrough`
 - Introduce new `LogTarget` for logging metrics to any `::Logger` compatible logger
+- Introduce new `OpenTelemetryTarget` for sending metrics via OpenTelemetry Metrics
 - Support for Puma 7
 
 ## [1.1.4] 2025-05-29